requirements.txt
flask
openai
requests
beautifulsoup4
matplotlib
textblob
scikit-learn
pandas
joblib
transformers
flask-cors
pymongo
python-dotenv
pyjwt
bcrypt